Fix numString type bug, detect redundant execution bug
The numString variable would in certain cases be converted to a
number-type but not converted back to a string-type. There is still
a bug where after inputting browser button inputs, if you press the
keyboard enter key it will execute the calculator function again
redundantly with no detected inputs. The extra execution does not
seem to change anything.